Don't duplicate code.

+++ update.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 build=0
+doInstall=1
 projects="Lightbox-Release Lightshow-Release"
 
 function usage
@@ -10,15 +11,18 @@ function usage
 	echo "Options:"
 	echo "  -b        Perform make"
 	echo "  -r        Perform full rebuild."
+	echo "  -o        Don't install; just do everything else (e.g. rebuild)."
 	echo "  -h        Print this message."
 }
 
-while getopts "brh" opt; do
+while getopts "brp:h" opt; do
 	case $opt in
 	b)
 		build=1;;
 	r)
 		build=2;;
+	o)
+		doInstall=0;;
 	p)
 		projects="$OPTARG";;
 	h)
@@ -54,24 +58,25 @@ if [ $build != 0 ]; then
 	done
 fi
 
-ssh root@192.168.69.2 "\
-	echo Mounting read/write...; \
-	mount -o remount,rw /lightbox; \
-	echo Stopping Chrome...; \
-	mv -f /lightbox/Chrome /lightbox/Chrome.old; \
-	echo Uploading Chrome...;"
+if [ $doInstall != 0 ]; then
+	ssh root@192.168.69.2 "\
+		echo Mounting read/write...; \
+		mount -o remount,rw /lightbox; \
+		echo Stopping Chrome...; \
+		mv -f /lightbox/Chrome /lightbox/Chrome.old; \
+		echo Uploading Chrome...;"
 
-ssh root@192.168.69.2 "cat > /lightbox/Chrome.new" < ../Lightshow-Release/built/DirectChrome
+	ssh root@192.168.69.2 "cat > /lightbox/Chrome.new" < ../Lightshow-Release/built/DirectChrome
 
-ssh root@192.168.69.2 "\
-	echo Fixing system...; \
-	[ -e /lightbox/fixture ] && sed s:fixture:scene:g </lightbox/fixture | sed s:output:driver:g >/lightbox/scene && rm -f /lightbox/fixture; \
-	[ -e /lightbox/scene ] && sed s:driver:fixture:g </lightbox/scene >/lightbox/scene.new && mv /lightbox/scene.new /lightbox/scene; \
-	echo Starting Chrome...; \
-	mv /lightbox/Chrome.new /lightbox/Chrome; \
-	chmod +x /lightbox/Chrome; \
-	echo Mounting read-only...; \
-	mount -o remount,ro /lightbox; \
-	echo Done."
-
+	ssh root@192.168.69.2 "\
+		echo Fixing system...; \
+		[ -e /lightbox/fixture ] && sed s:fixture:scene:g </lightbox/fixture | sed s:output:driver:g >/lightbox/scene && rm -f /lightbox/fixture; \
+		[ -e /lightbox/scene ] && sed s:driver:fixture:g </lightbox/scene >/lightbox/scene.new && mv /lightbox/scene.new /lightbox/scene; \
+		echo Starting Chrome...; \
+		mv /lightbox/Chrome.new /lightbox/Chrome; \
+		chmod +x /lightbox/Chrome; \
+		echo Mounting read-only...; \
+		mount -o remount,ro /lightbox; \
+		echo Done."
+fi
